+++ Demo/VComponentsDemo/Models/HomeRow.swift
@@ -15,11 +15,12 @@ enum HomeRow: Int, DemoableRow {
     case segmentedPicker, menuPicker, wheelPicker
     case stepper, slider,  rangeSlider
     case textField
-    case list, disclosureGroup
+    case sheet, disclosureGroup
+    case lazyScrollView, list, staticList
     case modal, bottomSheet, sideBar, alert, confirmationDialog, menu, contextMenu
     case toast
     case spinner, progressBar, pageIndicator
-    case text, sheet, lazyScrollView
+    case text
 
     // MARK: Properties
     static var sections: [DemoSection<HomeRow>] {
@@ -33,7 +34,7 @@ enum HomeRow: Int, DemoableRow {
             .init(id: 3, title: "Value Pickers", rows: [.stepper, .slider, .rangeSlider]),
             .init(id: 4, title: "Inputs", rows: [.textField]),
             .init(id: 5, title: "Containers", rows: [.sheet, .disclosureGroup]),
-            .init(id: 6, title: "Lists", rows: [.lazyScrollView, .list]),
+            .init(id: 6, title: "Lists", rows: [.lazyScrollView, .list, .staticList]),
             .init(id: 7, title: "Modals", rows: [.modal, .bottomSheet, .sideBar, .alert, .confirmationDialog, .menu, .contextMenu]),
             .init(id: 8, title: "Messages", rows: [.toast]),
             .init(id: 9, title: "Indicators", rows: [.spinner, .progressBar, .pageIndicator]),
@@ -69,6 +70,7 @@ enum HomeRow: Int, DemoableRow {
             
         case .lazyScrollView: return VLazyScrollViewDemoView.navBarTitle
         case .list: return VListDemoView.navBarTitle
+        case .staticList: return VStaticListDemoView.navBarTitle
 
         case .modal: return VModalDemoView.navBarTitle
         case .bottomSheet: return VBottomSheetDemoView.navBarTitle
@@ -116,6 +118,7 @@ enum HomeRow: Int, DemoableRow {
             
         case .lazyScrollView: VLazyScrollViewDemoView()
         case .list: VListDemoView()
+        case .staticList: VStaticListDemoView()
 
         case .modal: VModalDemoView()
         case .bottomSheet: VBottomSheetDemoView()