+++ Sources/VComponents/Components/Modals/Bottom Sheet/VBottomSheetUIModel.swift
@@ -83,7 +83,9 @@ public struct VBottomSheetUIModel {
         /// Spacing between label and close button. Defaults to `10`.
         public var labelCloseButtonSpacing: CGFloat = modalReference.layout.labelCloseButtonSpacing
         
-        /// Divider height. Defaults to `2/3`.
+        /// Divider height. Defaults to `0.67`.
+        ///
+        /// To hide divider, set to `0`, and remove header.
         public var dividerHeight: CGFloat = modalReference.layout.dividerHeight
     
         /// Divider margins. Defaults to `.zero`.